Split GlobalNamespaceProvider into separate files
