+++ SharpGen/BuiltinType.cs
@@ -1,0 +1,9 @@
+namespace SharpGen
+{
+    public enum BuiltinType
+    {
+        Marshal,
+        Math,
+        Unsafe
+    }
+}
